cocoshop/requirements.txt
asgiref==3.8.1
Django==4.2.1
djangorestframework==3.14.0
psycopg2==2.9.9
sqlparse==0.5.1
tzdata==2024.2